rename vis and layout functions
