+++ examples/16b_heterogeneity_multiple_ws_wd.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from floris.tools import FlorisInterface
-from floris.tools.visualization import visualize_cut_plane
+from floris.tools.flow_visualization import visualize_cut_plane
 
 
 """
