+++ examples/23_layout_visualizations.py
@@ -2,9 +2,9 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-import floris.tools.layout_functions as lf
+import floris.tools.layout_visulalization as lf
 from floris.tools import FlorisInterface
-from floris.tools.visualization import visualize_cut_plane
+from floris.tools.flow_visualization import visualize_cut_plane
 
 
 """
